+++ src/Proxy/TriangleProxy.php
@@ -136,6 +136,15 @@ class TriangleProxy extends Triangle implements ProxyInterface
         return $this->proxyGeometry->project($projector);
     }
 
+    public function rings() : array
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->rings();
+    }
+
     public function exteriorRing() : \Brick\Geo\LineString
     {
         if ($this->proxyGeometry === null) {